+++ TestSutraUsers/src/main/java/com/tathvatech/workstation/oid/UnitInProjectOID.java
@@ -1,9 +1,10 @@
 package com.tathvatech.workstation.oid;
 
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
-public class UnitInProjectOID extends OID{
+import com.tathvatech.user.OID.OID;
+
+public class UnitInProjectOID extends OID {
 
 	public UnitInProjectOID(int pk)
 	{
@@ -37,7 +38,7 @@ public class UnitInProjectOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
